move getValue to primitives components
