+++ packages/tc-ui/src/pages/edit/template.jsx
@@ -1,46 +1,9 @@
 const React = require('react');
-const { sortBy } = require('lodash');
 const { getEnums } = require('@financial-times/tc-schema-sdk');
 const { FormError } = require('../../lib/components/messages');
 const { Concept, SectionHeader } = require('../../lib/components/structure');
-
+const { getValue } = require('../../primitives/server');
 const { SaveButton, CancelButton } = require('../../lib/components/buttons');
-
-const getValue = (itemSchema, itemValue) => {
-	// preserves boolean values to prevent false being coerced to empty string
-	if (itemSchema.type === 'Boolean') {
-		return typeof itemValue === 'boolean' ? itemValue : '';
-	}
-
-	// return relationships as type, code and name object
-	if (itemSchema.relationship) {
-		if (itemSchema.hasMany) {
-			return itemValue
-				? sortBy(itemValue, `${itemSchema.type}.code`).map(item => ({
-						code: item.code || item[itemSchema.type].code,
-						name:
-							item.name ||
-							item.code ||
-							item[itemSchema.type].name ||
-							item[itemSchema.type].code,
-				  }))
-				: [];
-		}
-		return itemValue
-			? {
-					code: itemValue.code || itemValue[itemSchema.type].code,
-					name:
-						itemValue.name ||
-						itemValue.code ||
-						itemValue[itemSchema.type].name ||
-						itemValue[itemSchema.type].code,
-			  }
-			: null;
-	}
-
-	// everything else is just text
-	return itemValue;
-};
 
 const PropertyInputs = ({ fields, data, isEdit, type, assignComponent }) => {
 	const propertyfields = Object.entries(fields);
@@ -67,9 +30,13 @@ const PropertyInputs = ({ fields, data, isEdit, type, assignComponent }) => {
 				lockedBy = fieldsToLock[name];
 			}
 			const { EditComponent } = assignComponent(item);
+			const itemValue = item.isRelationship
+				? data[`${name}_rel`]
+				: data[name];
+
 			const viewModel = {
 				propertyName: name,
-				value: getValue(item, data[name] || data[`${name}_rel`]),
+				value: getValue(item, itemValue),
 				dataType: item.type,
 				parentType: type,
 				options: getEnums()[item.type]
